# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MTAA Fitness API",
        "description": "API for user registration, login, workouts, friends and plans.",
        "contact": {
            "email": "support@fitness-app.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/plan/add": {
            "post": {
                "tags": [
                    "Workout Plan"
                ],
                "summary": "Add a workout to user's plan",
                "operationId": "274ade3127198acfffe2b90c3d5b9f72",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workout_id",
                                    "repetitions",
                                    "sets"
                                ],
                                "properties": {
                                    "workout_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "repetitions": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "sets": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Workout added to your plan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/plan/update": {
            "put": {
                "tags": [
                    "Workout Plan"
                ],
                "summary": "Update repetitions and sets in the user's plan",
                "operationId": "4552c2f208c37548e8de3fd830731633",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workout_id",
                                    "repetitions",
                                    "sets"
                                ],
                                "properties": {
                                    "workout_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "repetitions": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "sets": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Workout plan updated successfully"
                    },
                    "404": {
                        "description": "Workout not found in plan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/plan": {
            "get": {
                "tags": [
                    "Workout Plan"
                ],
                "summary": "Get user's workout plan",
                "operationId": "6a7bfe079f4f3128195ed3b199f5db3d",
                "responses": {
                    "200": {
                        "description": "List of workouts in user's plan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/plan/remove/{workout_id}": {
            "delete": {
                "tags": [
                    "Workout Plan"
                ],
                "summary": "Remove workout from user's plan",
                "operationId": "8037af73aa3dec9bbea3b115b01e36d8",
                "parameters": [
                    {
                        "name": "workout_id",
                        "in": "path",
                        "description": "Workout ID to remove",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workout removed from plan"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}